pyproject.toml
[tool.poetry]
name = "sec-agent-v2"
version = "0.1.0"
description = ""
authors = ["Karim Mardhani <kmardhani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
openai = "^1.25.1"
python-dotenv = "^1.0.1"
llama-index = "^0.10.34"
sec-api = "^1.0.18"
pyautogen = "^0.2.27"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
